Move classes into packages by functionality
This adds export and display packages, and moves a couple parsing
classes into xfl. Eventually xfl and display will be split out into
their own swc as a runtime for starling, but I'll need to wean them off
of aspirin to do that.

Also, rename the export classes that make files to Publisher. Exporters
drives the process, Publishers make files.