Screen wipe test XFL.
I've screwed around a lot with it at this point, but for this commit I've left one important layer
off guide: the first block_b_lg layer shows the case where a layer has frames that have no display
objects.

block_r_lg could be used for color alpha testing, that layer's modifications are an attempt to make
sure there always is an object, but that it's just alpha = 0 when there should be nothing. Finally,
the Border layer shows that this tool flump doesn't support scale-9 grids, but I'm not even sure
if that's something we easily handle here. Perhaps a feature for down the road.