Only complete xfl loading when parsing is complete
Previously it'd be marked as complete if the swf and xml loading
executors were both shutdown when one of them completed its work. The
swf loader only had a single job, and it shutdown when that job
completed. The xml loading had as many jobs are there are symbols, and
it shutdown when all those jobs were submitted.  If the swf parsing
completed before the xml parsing, this meant the whole thing would be
marked completed while xml was still being parsed and the library would
remain in a weird half-resolved state. Man that was a pain to track
down.