require that a symbol's export name match its library name
A mismatched texture export name causes an error in betwixt
(textures are named by their export name, but referenced by movies by
their library names). This could probably be fixed in betwixt, but this
makes things clearer for everyone anyway (referenced movies must be
referenced by library name, because they may not be exported; it would
be weird to have different naming rules for exported textures and
movies).

It might make sense to replace "symbol" with an "exported" flag (and
rename libraryItemName to symbol).