Remove library names from exported data
XflLibrary now does a second pass on the parsed keyframes and replaces
their libraryName reference with the symbol or implicit symbol that
libraryName points to. This means libraryNames can now be left out of
the export altogether, and we no longer need to require matching library
names and symbols.

I believe betwixt will work as is on the xml export, but flambe is going
to need a couple changes. The JSON for AtlasTextureMold now uses symbol
instead of name as a key, and MovieMold uses id instead of symbol.
I think updating to use those new keys should be all that's required.