Combine texture origins into keyframe data
This allows us to skip the "holder" Sprite that wrapped every
instantiated Image, which means that a Flump runtime display list will
be identical in structure to its corresponding Flash display list.

I also changed the name and inverted the value of TextureMold.offset -
it's now TextureMold.origin.

THIS IS A BREAKING CHANGE, so I've updated the data version to 2.