Added an option to generate mipmaps for textures.
It makes loading slower and there is a slight memory penalty, but textures that use a large variety
of scales look a hell of a lot better with mipmapping enabled. trilinear smoothing helps a lot too
but the end flump user can turn that on himself by traversing the loaded Movie's Image children
and changing the smoothing property.