Expose the base Library implmentation.
My project's frankenweenie has gotten enough out of hand that I've ended up writing my own entire
loader. Having a handy library implementation to use is useful though.

The loader I've written caches the unzipped data to a local SharedObject and can read it back
out later. Some of the implementation details are specific to my project, but a similar loader could
probably be spun up, or attached as an option to the main loader, if someone else wanted the
functionality. I might just throw my entire loader stuff into a gist or something for people to
make use of if they want it.

The process of unzipping the flump zip and decoding the PNGs is slow enough that caching the
fully decoded data locally is a flump loading time win for my project.