Solve the "empty layer frames" issue in the runtime, not the exporter
Instead of adding empty keyframes to Layers at export time, we do the
right thing at runtime and swap in an empty sprite when a layer is asked
to draw a frame past the end of its frame count.

This is closer in spirit to what Flash does, and doesn't require assets
to be re-exported.

(It does require that other runtimes be updated in order to take
advantage of the fix, of course.)