Scaffolding for exporting from the command line.
I had originally assumed I'd want to package the bash script up with the .air app, but Tim's
convinced me that it makes more sense to keep it separate. My initial concern was that the script
will likely be necessarily tied to changes in the arguments the exporter expects, but we'll just
need to tag the script for releases like we do the .air.