Wait a frame before doing actual publish.
Publishing always involves a very long frame for any sizable flump project, so let's start it on
a new frame. For one thing, it appears that the output only flushes between frames, and it looks
better from the CL script you've been told that publishing is started before the app halts for
awhile.